revert float32 after reshape

+++ gptqmodel/quantization/gptq.py
@@ -235,7 +235,7 @@ class GPTQ:
             self.process_batch(inp)
 
     def process_batch(self, inp: torch.Tensor):
-        reshaped_inp = inp.to(device=DEVICE_1)
+        reshaped_inp = inp.to(device=DEVICE_1, dtype=torch.float32)
         del inp
 
         # input reshaping
@@ -252,7 +252,6 @@ class GPTQ:
             reshaped_inp = unfold(reshaped_inp)
             reshaped_inp = reshaped_inp.transpose(1, 2).flatten(0, 1)
 
-        reshaped_inp = reshaped_inp.to(dtype=torch.float32)
         batch_token_size = reshaped_inp.shape[0]
 
         if self.H is None:
